fixing the bug

+++ lib/RunningAveragedB/src/RunningAveragedB.cpp
@@ -1,38 +1,16 @@
 // RunningAveragedB
 // dmf 5.22.20 
 // generate a 'correct' average of dB measurements 
-//
-
-//
-//    FILE: RunningAverage.cpp
-//  AUTHOR: Rob Tillaart
-// VERSION: 0.2.06
-//    DATE: 2015-mar-07
-// PURPOSE: RunningAverage library for Arduino
-//
-// The library stores N individual values in a circular buffer,
-// to calculate the running average.
-//
-// HISTORY:
-// 0.1.00 - 2011-01-30 initial version
-// 0.1.01 - 2011-02-28 fixed missing destructor in .h
-// 0.2.00 - 2012-??-?? Yuval Naveh added trimValue (found on web)
-//          http://stromputer.googlecode.com/svn-history/r74/trunk/Arduino/Libraries/RunningAverage/RunningAverage.cpp
-// 0.2.01 - 2012-11-21 refactored
-// 0.2.02 - 2012-12-30 refactored trimValue -> fillValue
-// 0.2.03 - 2013-11-31 getElement
-// 0.2.04 - 2014-07-03 added memory protection
-// 0.2.05 - 2014-12-16 changed float -> double
-// 0.2.06 - 2015-03-07 all size uint8_t
-//
-// Released to the public domain
+// The running average is limited to uint_8 (256) 
+// values, so is used only to maintain an ongoing 
+// 'ambient' value in the SoundMeter.ino code
 //
 
 #include "RunningAveragedB.h"
 #include <stdlib.h>
 #include <math.h>
 
-RunningAveragedB::RunningAveragedB(uint16_t size)
+RunningAveragedB::RunningAveragedB(uint8_t size)
 {
     _size = size;
     _ar = (double*) malloc(_size * sizeof(double));
